requirements.txt
# Web 框架
Flask==3.1.1
Flask-SocketIO==5.3.3
flask-cors==6.0.1
python-socketio==5.13.0
python-engineio==4.12.2
Werkzeug==3.1.3

# 音频处理
PyAudio==0.2.14
numpy==2.3.2
funasr==1.2.6

# 向量搜索和嵌入
faiss-cpu==1.11.0.post1
sentence-transformers==5.0.0
text2vec==1.2.0
scikit-learn==1.7.1

# 深度学习
huggingface-hub==0.34.3

# PDF 处理
PyMuPDF==1.21.1
fitz==0.0.1.dev2

# NLP处理
spacy==3.8.7

# 工具和辅助库
requests==2.32.4
tqdm==4.67.1
python-dotenv==1.1.1
loguru==0.7.3